refactor: pequenos ajustes nso cards, chip, e planet
