+++ resources/js/Components/Atoms/Planet.jsx
@@ -3,10 +3,10 @@ import { styled } from "@mui/material/styles";
 
 const PlanetStyle = styled('img')(({ theme }) => ({
     boxShadow: '-4px 4px 8px rgba(26, 27, 35, 0.15)',
-    borderRadius: '24px',
+    borderRadius: '50%',
     width: '100%',
-    height: '100%',
-    objectFit: 'cover',
+    height: 'auto',
+    objectFit: 'cover'
 }));
 
 const Planet = ({ src, alt }) => {
